Update Lucene to 9.3.0

+++ server/src/testFixtures/java/org/apache/lucene/index/PatchedAssertingLeafReader.java
@@ -1029,7 +1029,7 @@ public class PatchedAssertingLeafReader extends FilterLeafReader {
         }
 
         @Override
-        public long docValueCount() {
+        public int docValueCount() {
             return in.docValueCount();
         }
